helper: fix get_array_expr()
The problem here is that say you have a statement like:

	foo[x] = whatever();

When we're parsing the assignment, the parent of foo[x] hasn't been
initialized yet, so get_array_expr() returns NULL.  Then later it does
get initialized and after that it returns the correct thing.  So we end
up creating two states in smatch_extra which aren't in sync.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>