implied: add a timeout in smatch_implied.c
If we go over a minute in parsing a function then stop trying to handle
implications.

The story here is that blk_stack_limits() was taking 6 minutes on my
system and I thought this would help.  Except when I tested it, the last
time we called this function then it was at the 58 second mark so it didn't
help at all.

But it feels right so I'm going to do it.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>