build: remove unused support for pkgconfig
Sparse being a library was installed as such and with support for
pkgconfig.

However, it seems that sparse is only used for its tools and not
as a library (at least not as an external one) and currently
doesn't offer a stable interface.

For now, remove the support for this, it will be easy enough to
restore it if there is a new need.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>