constraints: don't store signed vs unsigned
In the DB, we don't store the sign vs unsigned and it's sort of complicated
to add it.  And this is array index stuff and we have a separate check
which looks for underflows so that's a bit covered already.

So let's strip out the signed vs unsigned bits so that if a comparison is
unsigned, it still fits the requirements.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>