IR: let .cond unionize with .src and not .target
In struct instruction, .target is normally used to hold
the result. Its value is thus produced/defined by instructions.
On the contrary, .cond is used as an input value and is
thus used by instructions.

However, these two fields belong to the same union. This
creates slight complications for code, like liveness analysis
which care about which fields are used and which are defined
by the instructions.

Change this by unionizing .cond with .src, .src1 & friends instead
of with .target.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>