implications: set_state() should trump implications
The problem here is that say it's implied that "a <= b" but the condition
is "a < b".  The final true state should be "a < b" but it was instead set
to the implied state.

The exception is for smatch_extra states where the implied state and the
condition are always going to be the same, but the implied state has
implications associated with it which we want to preserve.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>