struct_assignment: fix handling of non_struct_assignments
We already did a remove_addr() so it means that we always treat memcpy()
as setting the destination to unknown.

With the modified code, if you don't have the cross function DB built and
it's parsing code like:

	memcpy(&new_int, &old_int, sizeof(new_int));

Then it still doesn't work because it sees that we're passing an address
and clears the value.  Also the src pointer is const so that shouldn't
happen.  Oh well, bugs on top of bugs.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>