db/mtags: add mtag_data and mtag_map tables
The mtag_data table stores information about the data stored in the
mtag.  In other words it stores that $->bar is 42.

The mtag_map is way to record that a pointer is stored inside another
struct.  The offset values are negative because it stores the
container_of() pointer.  Normally, in C we have things like "foo->bar->baz"
where if you start with "foo" then you can find "baz".  But this is the
reverse where you start with "baz" and find "bar" then find "foo".

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>