testsuite: add support for 'format -a'
The 'format' command create the information needed for the testcase
from the input file and output this on stdout. The developper must
then add this to the input file.

Let's do this automatically by adding an option '-a' to the 'format'
command to directly append the infos to the input file.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>