param_set/filter: PARAM_SET instead of PARAM_ADD
The difference between PARAM_SET and PARAM_ADD is that PARAM_SET doesn't
save the values that we passed in.  It uses a call to
param_has_filter_data() to see if it should be PARAM_SET or PARAM_ADD.

The bug is that param_has_filter_data() assumes that if a parameter has not
been set then it means it has filter data.  But param_filter doesn't save
everything.  So this is a hack but it should work for most things which is
that we look up if a parameter has been assigned and if so then that means
it doesn't have filter data.

There are possibly other places with no filter_data which should still be
PARAM_SET but this is at least an improvement...

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>