extra: pass the modified expression to extra_mod_hooks (if easy)
It's probably more ideal to pass the modified expression to extra_mod_hooks
than to pass the name/sym.  The problem is that for the DB then we don't
have a expression.  The solution is to pass what we have.  It is a lazy and
less than ideal way to go.  But it's sort of the Smatch philosophy to use
the information that we have instead of bemoaning that everything is not
perfect...

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>