helper: keep a small cache of recently used variable names
Smatch does a lot of changing expression to str/sym pairs.  It ends up
that we call expr_to_str(expr) over and over for the same expression.
So instead of doing the work everytime, we can keep a small cache of the
most recently used variables to speed up the process.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>