testsuite: reset 'quiet' at the start of each testcase
The flag 'quiet' is used to quiets unwanted error messages,
for example for testcases known to fail, but this flag is reset
too late so that the beginning of the next testcases will run
with the value for the previous case.

Fix this by reseting the flag at the begining of each testcase.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>