fix '__SIZE_TYPE__' for LLP64
size_t_ctype is set to uint, ulong or ullong, depending on
the architecture (ullong is only used for LLP64).

However, when emitting '__SIZE_TYPE__', it's only compared to ulong
or uint.

Fix this by using an small helper directly using the right
struct symbol * and using builtin_typename() to output the
right type. This way we're guaranteed that '__SIZE_TYPE__'
is kept coherent with the internal type: size_t_ctype.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>