fix: spaces in macro definition on the command line
GCC's manual or POSIX say about the '-D' option something like:
    '−D name[=value]' should be treated as if in directive
    '#define name value' (with '1' as default for the value),
    including its tokenization.
So an option like '-DM(X, Y)=...' should be processed like a
directive '#define M(X, Y) ...'.

However, the current code treat a space as a separator between
the macro and its definition, just like the '='. As consequence,
the above option is processed like the directive would be
'#define M(X, Y)=...', with 'M(X,' as the macro (name) and
'Y)=...' as its definition.

Fix this by stopping to treat the space character specially,
thus only using '=' as the separator.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>