fix: accept 'sparse -D M...'
Till now, sparse was unneedlessly strict in what it accepted in
'-D' options. More specifically, it doesn't accept:
1) separated '-D' and the macro definition, like:
	sparse -D MACRO[=definition] ...
2) a space between the '-D' and the macro name, like:
	sparse '-D MACRO[=definition] ...

Case 1) is clearly accepted by GCC, clang and should be
accepted for a POSIX's c99. Case 2's status is less clear
but is also accepted by GCC and clang (leaving any validation
to the corresponding internal #define).

Fix this by accepting separated command line argument for '-D'
and the macro (and removing the check that rejected the macro
part if it started with a space).

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>