testsuite: allow default args from environment for test commands
During testing it's sometimes useful to force some default arguments
for all commands. An example of this is using '-m32' which essentially
allow to run the tessuite on an 64bit machine as-if run a 32-bit one.

Allow this by using the environment variable 'SPARSE_TEST_ARGS' to
hole default arguments for the test commands.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>