llvm: add declares for function prototypes
Currently, for the LLVM's IR, the functions are only declared
when they are called (or otherwise used) or when they are defined,
whichever come first. But systematically adding a declaration
for each function we have the prototype shouldn't hurt and
would add one level of type checking.

Change this by adding the declaration (via get_sym_value())
for each prototype we have.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>