llvm: simplify emit of null pointers
Most pointers with a constant value are simply null-pointers.
The only exception is when a known address is casted to pointer.

The current code only handle code for the general case:
emit code for a constant integer and then cast this to a pointer.
This obfuscate a bit the ouput, making it hard to read.

Change this by special handling the normal case of null-pointers
by directly using LLVM's LLVMConstPointerNull().

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>