llvm: default init of arrays & structs
Linearization of arrays & structs default initialization is done
as if these objects would be an integer as large as the object.
This integer is then simply zero-initialized.
(This may be objectionable and will be done differently).

However, sparse-llvm is not ready to handle this situation where
a non-scalar is initialized with a scalar value.

Workaround this by using LLVMConstNull() when possible.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>