ptrlist: let FOR_EACH_PTR() ignore tags
Since the introduction of the macro FOR_EACH_PTR_TAG() all the users
of the macros FOR_EACH_PTR() & FOR_EACH_PTR_REVERSE() are only used
for the usual case where no tag is used on the pointers.

It is thus no more needed to strip the tag from these pointers
and internally we can, for these two macros, call the simpler
PTR_ENTRY_NOTAG() instead of PTR_ENTRY_UNTAG().

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>