links: no need to record links outside a function call
Storing the links ends up being not very scalable.  The whole idea of
links is probably not the right thing...  How this shows up in the kernel
is parsing huge arrays takes a very long time.  Most of these are global.
And it turns out we don't care about links at all for globals so it's a
special case and we can hack around the problem instead of figuring out a
complete fix.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>