mtag: introduce create_mtag_alias()
The idea here is that we have a many to one or one to many relationship so
we create aliases for the many.  So for example, you have a function which
saves a parameter to a mtag+offset then you give you generate an mtag and
an alias for your parameter.

The alias doesn't have to be persistent across builds but it helps for
debugging.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>