address: fix handling of &foo->bar
The old code was pretty rubbish.  But say you have &foo->bar and ->bar
isn't the first member of the struct, then you "know" that &foo->bar is
non-NULL.  In theory it could be NULL, but if that's true you're already
screwed so it's not worth worrying about.

This code tries to implement that.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>