constraints_required: improve handling of assigned sizes
This is tricky because I'm trying to remember what was going on before
Christmas.  It has to do with code like this:

	size = num * sizeof(*foo);
	p = alloc(size);

What this code does is it uses both "num * sizeof(*foo)" and "size" as
the size.  I'm not sure why we would want to do that, but presumably there
was a reason.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>