check_atomic_inc_dec: track atomic_inc() and atomic_dec()
I had some ideas that I would find leaked atomic_inc() overflows with this
where we increment a counter to show that something is in use but forget
to decrement it on some error paths.

That project turned out harder to implement than I had hoped.  But I still
want to commit this because I use the information for other checks.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>