function_hooks: fake an assignment when functions return "0-s32max[$0->bar]"
Say a function returns a parameter or a member of a parameter like this:

struct my_type *get_bar(struct foo *foo)
{
	return foo->bar;
}

Then we can create a fake assignment for the caller:  my_bar = foo->bar;

The reason I didn't do this before was that it was a bit complicated to
take a string and generate an expression.  It turns out that when you
call a function, then you've already got most of the expression, you just
need to add the "->bar" part which isn't so hard.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>