implied: use a time based timeout instead of counting ->nr_children
I've been running this code for a while.  It seems to work.  It's not
beautiful or complete.  I'm sorry.  I want to commit it now so that I make
my diff smaller.

The problem is that ->nr_children is not really the right thing to measure.
We could have a huge number of children but only need to look at 20 to
decide how the implications fit together.  The effect of this change is
hopefully to make implications more accurate by using some implications
which we would have ignored before.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>