add a field 'tainted' to struct instruction
During the processing, some verification are done and if the
check fails, a warning is normally issued. But some checks are
done at several steps and so the same warning can be give
several time, which is annoying.

Add to instructions a field '.tainted' which will allow to
mark instructions which fail some checks and have already
warned about.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>