flow: set loop_count to zero when parsing inline functions
I deliberately didn't do this before because it's a trade off.  The choice
is either mark some returns as impossible when they are actually possible
on the second iteration through the loop or mark some returns as possible
when they are impossible.  Either way is going to cause some confusion.

I have decided to mark returns as impossible.  This does cause a couple
issues where some unpublished code complains that "why are we checking
for IS_ERR() when that's not possible".  I will write a work around for
that eventually.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>