constraints: handle conditions like "if (8 < x) {"
All right thinking people write code with the variable on the left and the
limit on the right.  That's how people talk in English.  Maybe in other
languages people say "if the limit is less than the variable"?  Who knows.

Anyway, I sometimes like to ignore wrong headed code, but I always have to
go back and change it to handle flipped logic as well and that's what this
patch does.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>