fix examination of bad typeof
When the operand of typeof() is invalid, the corresponding
type is also invalid and can't be used. But currently,
when the corresponding SYM_TYPEOF symbol is examined,
this symbol is then simply returned as is. This, of course,
create problem for subsequent code since an examined type is
not supposed to be a SYM_TYPEOF anymore (one of the symptoms
will be warning about "unknown type 11").

Fix this by changing the SYM_TYPEOF into a SYM_NODE (as it's
expect) but pointing to bad_ctype. So further processing gently
continue and the 'bad_ctype' will do its job when needed.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>