remove warning "call with no type"
This warning is issued when trying to linearize a call
expression with a null expr->ctype. However, such null
ctypes are a consequence of earlier errors detected
during evaluation and for which a warning, specific
about the nature of the problem, have already been issued.

In short, this "call with no type" is non-informative
and redundant, so avoid useless noise and remove this
warning.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>