early return if null ctype in evaluate_conditional()
This function contains a few tests which must only be done
if the type is valid, same for the final call to degenerate().

Currently this is done by an "if (ctype) { ... }" but this
cause multiple NULL tests, take screen real estate and is
somehow error prone.

Change this by returning as soon as we know the ctype is
invalid since nothing good can be done after it.
This also makes, IMO, the code clearer.

NB: no functional changes here.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>