use show_pseudo() for OP_SYMADDR's symbol
In show_instruction(), what is done to display OP_SYMADDR's
symbol is very close to what is done to display a PSEUDO_SYM.
In fact, what is done by show_pseudo() is more complete and
more informative.

So simply use show_pseudo() to display this part. This will
also avoid crashes caused by a NULL symbol.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>