extra: cleanup handling "if (foo & 0x1) {"
These two chunks are basically the same.  I've moved them to their own
function to fix cut down on the duplicate code...  And, oh, I spot a copy
and paste bug in the first block where I say "expr->left" instead of
"expr->right", so it's a bug fix as well.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>