OP_CALL should use the full function type
OP_CALL keep a list with the type of the function itself
and of each of its arguments.

However, the function type added to the list is not the complete type
but its base type. So, we can't use the function's modifiers or contexts.

Fix this by storing the complete function type.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>