use function-like syntax for __range__
One of sparse's extension to the C language is an operator
to check ranges. This operator takes 3 operands: the expression
to be checked and the bounds.

The syntax for this operator is such that the operands need to
be a 3-items comma separated expression. This is a bit weird
and doesn't play along very well with macros, for example.

Change the syntax to a 3-arguments function-like operator.

NB. Of course, this will break all existing uses of this
    extension not using parenthesis around the comma
    expression but there doesn't seems to be any.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>