move the optimization loop in its own file
Once linearized, a few essentials optimizations are done
but the logic of the succession of these optimizations
is open-coded at the end of linearize_fn().

Obviously, this logic have nothing to do with the
linearization and moving it to a separate file will help
to clarify this logic and change it when needed.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>