build: add -MP for generated dependencies
When an header file is removed, or more often when
bisecting or more generaly when we checkout an older
version where some header file didn't exist yet,
we typically have generated .d files containing
dependencies on a non-existing header file. In this
case, make abort and complains about the dependencies
that can't be met.

Avoid this situation by using the '-MP' option so that GCC
can automatically add a dummy rule for those header file.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>