move liveness interface to its own header
Currently, the interface for liveness analysis is declared
in the flow.h header file. There is no a real need to move
it to its own file but it's the dual of the previous patch
and help a bit to make the code structure clearer.

Move the prototype for liveness analysis to its own header.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>